requirements.txt
pandas>=2.0.0
yfinance>=0.2.0
numpy>=1.24.0
matplotlib>=3.7.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
azure-functions>=1.18.0
azure-storage-blob>=12.19.0

